+++ network/src/control.rs
@@ -49,7 +49,7 @@ pub trait Control: Send + Sync {
     fn get_whitelist(&self) -> Result<(Vec<FilterEntry>, bool), Error>;
     fn get_blacklist(&self) -> Result<(Vec<FilterEntry>, bool), Error>;
 
-    fn recent_network_usage(&self) -> Result<HashMap<&'static str, usize>, Error>;
+    fn recent_network_usage(&self) -> Result<HashMap<String, usize>, Error>;
 }
 
 #[derive(Clone, Debug)]