+++ sync/src/block/downloader/header.rs
@@ -74,9 +74,18 @@ impl HeaderDownloader {
     }
 
     pub fn update(&mut self, total_score: U256, best_hash: H256) -> bool {
-        if self.total_score < total_score {
+        if self.total_score == total_score {
+            true
+        } else if self.total_score < total_score {
             self.total_score = total_score;
             self.best_hash = best_hash;
+
+            if self.client.block_header(&BlockId::Hash(best_hash)).is_some() {
+                self.pivot = Pivot {
+                    hash: best_hash,
+                    total_score,
+                }
+            }
             true
         } else {
             false
@@ -98,6 +107,10 @@ impl HeaderDownloader {
             Some(header) => header.clone(),
             None => self.client.block_header(&BlockId::Hash(self.pivot.hash)).unwrap(),
         }
+    }
+
+    pub fn pivot_score(&self) -> U256 {
+        self.pivot.total_score
     }
 
     pub fn is_idle(&self) -> bool {
@@ -125,7 +138,11 @@ impl HeaderDownloader {
     /// Expects importing headers matches requested header
     pub fn import_headers(&mut self, headers: &[Header]) {
         let first_header_hash = headers.first().expect("First header must exist").hash();
-        if first_header_hash == self.pivot.hash {
+
+        // This happens when best_hash is imported by other peer.
+        if self.best_hash == self.pivot.hash {
+            ctrace!(SYNC, "Ignore received headers, pivot already reached the best hash");
+        } else if first_header_hash == self.pivot.hash {
             for header in headers.iter() {
                 self.downloaded.insert(header.hash(), header.clone());
             }
@@ -157,6 +174,13 @@ impl HeaderDownloader {
     pub fn mark_as_imported(&mut self, hashes: Vec<H256>) {
         for hash in hashes {
             self.downloaded.remove(&hash);
+
+            if self.best_hash == hash {
+                self.pivot = Pivot {
+                    hash,
+                    total_score: self.total_score,
+                }
+            }
         }
     }
 }