+++ test/src/helper/spawn.ts
@@ -65,6 +65,7 @@ export default class CodeChain {
     private readonly _chain: ChainType;
     private readonly _rpcPort: number;
     private readonly argv: string[];
+    private readonly env: { [key: string]: string };
     private isTestFailed: boolean;
     private process?: ChildProcess;
     private readonly keyFileMovePromise?: Promise<{}>;
@@ -112,9 +113,10 @@ export default class CodeChain {
             argv?: string[];
             additionalKeysPath?: string;
             rpcPort?: number;
+            env?: { [key: string]: string };
         } = {}
     ) {
-        const { chain, argv, additionalKeysPath } = options;
+        const { chain, argv, additionalKeysPath, env } = options;
         this._id = CodeChain.idCounter++;
 
         const { rpcPort = 8081 + this.id } = options;
@@ -124,7 +126,9 @@ export default class CodeChain {
         mkdirp.sync(`${projectRoot}/keys/`);
         mkdirp.sync(`${projectRoot}/test/log/`);
         this._dbPath = mkdtempSync(`${projectRoot}/db/`);
-        this._ipcPath = `/tmp/jsonrpc.${this.id}.ipc`;
+        this._ipcPath = `/tmp/jsonrpc.${new Date()
+            .toISOString()
+            .replace(/[-:.]/g, "_")}.${this.id}.ipc`;
         this._keysPath = mkdtempSync(`${projectRoot}/keys/`);
         if (additionalKeysPath) {
             this.keyFileMovePromise = new Promise((resolve, reject) => {
@@ -146,14 +150,22 @@ export default class CodeChain {
         this._sdk = new SDK({ server: `http://localhost:${this.rpcPort}` });
         this._chain = chain || "solo";
         this.argv = argv || [];
+        this.env = env || {};
         this.isTestFailed = false;
     }
 
-    public async start(
-        argv: string[] = [],
-        logLevel = "trace,mio=warn,tokio=warn,hyper=warn,timer=warn",
-        disableLog = false
-    ) {
+    public async start(params?: {
+        argv?: string[];
+        logLevel?: string;
+        disableLog?: boolean;
+        disableIpc?: boolean;
+    }) {
+        const {
+            argv = [],
+            logLevel = "trace,mio=warn,tokio=warn,hyper=warn,timer=warn",
+            disableLog = false,
+            disableIpc = true
+        } = params || {};
         if (this.keyFileMovePromise) {
             await this.keyFileMovePromise;
         }
@@ -162,18 +174,24 @@ export default class CodeChain {
         // NOTE: https://github.com/CodeChain-io/codechain/issues/348
         process.env.WAIT_BEFORE_SHUTDOWN = "0";
 
+        const baseArgs = [...this.argv, ...argv];
+        if (disableIpc) {
+            baseArgs.push("--no-ipc");
+        } else {
+            baseArgs.push("--ipc-path");
+            baseArgs.push(this.ipcPath);
+        }
+
         // Resolves when CodeChain initialization completed.
         return new Promise((resolve, reject) => {
             this.process = spawn(
                 `target/${useDebugBuild ? "debug" : "release"}/codechain`,
                 [
-                    ...this.argv,
-                    ...argv,
+                    ...baseArgs,
                     "--chain",
                     this.chain,
                     "--db-path",
                     this.dbPath,
-                    "--no-ipc",
                     "--keys-path",
                     this.keysPath,
                     "--no-ws",
@@ -186,7 +204,11 @@ export default class CodeChain {
                 ],
                 {
                     cwd: projectRoot,
-                    env: process.env
+                    env: {
+                        ...process.env,
+                        ENABLE_DELEGATIONS: "true",
+                        ...this.env
+                    }
                 }
             );